added 2nd order explanation figure
This explanation figure shows an example matrix denoting the regression of the phenotype onto 2 sites. In the test data, the file ending with "rFon2D" contains this matrix for this particular data (the 2nd matrix out of the 4 that are in the output). So, if we're looking at sites 3 and 5, this matrix shows us all the possibilities of nucleotide combinations at each site. One can visualize this as a heatmap or via another way.